client/client.c: Fixed format problem with %d -> %.0f. rpc_client/cli_srvsvc.c rpcclient/cmd_lsarpc.c rpcclient/cmd_reg.c rpcclient/cmd_samr.c: Fixed rpcclient compile problems introduced by the new RPC code. Jeremy.

+++ source/client/client.c
@@ -463,7 +463,7 @@ static void cmd_du(void)
 
 	do_dskattr();
 
-	DEBUG(0, ("Total number of bytes: %d\n", dir_total));
+	DEBUG(0, ("Total number of bytes: %.0f\n", (double)dir_total));
 }
